DAY_004_File_Uploader_Appv1/backend/requirements.txt

Flask
Flask-Cors
pytesseract
Pillow
